+++ R/PredictionDataCompRisks.R
@@ -1,5 +1,8 @@
 # Here we define some mlr3-mandatory S3 methods of the `PredictionDataSurv` object
 
+#' @rdname as_prediction_cmprsk
+#' @param check (`logical(1)`)\cr
+#'   Whether to validate internal `PredictionDataCompRisks` data during conversion with [mlr3::as_prediction()].
 #' @export
 as_prediction.PredictionDataCompRisks = function(x, check = FALSE, ...) {
   invoke(PredictionCompRisks$new, check = check, .args = x)